+++ test/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_09_15_010002) do
+ActiveRecord::Schema[8.1].define(version: 2026_09_15_010003) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
   enable_extension "pgcrypto"
@@ -152,6 +152,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_09_15_010002) do
 
   create_table "recording_studio_terms_and_conditions_terms", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.text "body", null: false
+    t.text "change_note"
     t.datetime "created_at", null: false
     t.string "title", null: false
   end